fix(travis): force extration of consul

+++ tests/travis/start_consul.sh
@@ -10,7 +10,12 @@ VERSION=0.6.1
 ARCHIVE_NAME=${NAME}_${VERSION}_linux_amd64.zip
 
 if [ ! -f ${NAME} ]; then
-  curl -O -insecure https://releases.hashicorp.com/${NAME}/${VERSION}/${ARCHIVE_NAME}
+  if [ ! -f ${ARCHIVE_NAME} ]; then
+    rm ${ARCHIVE_NAME}
+    curl -O -insecure https://releases.hashicorp.com/${NAME}/${VERSION}/${ARCHIVE_NAME}
+  fi
+
+  rm ${NAME}
   unzip -o ${ARCHIVE_NAME}
 fi
